+++ packages/convert-svg-core/src/Converter.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 Alasdair Mercer
+ * Copyright (C) 2022 neocotic
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,6 +24,7 @@
 
 /* global document: false */
 
+const cheerio = require('cheerio');
 const fileUrl = require('file-url');
 const fs = require('fs');
 const path = require('path');
@@ -44,6 +45,8 @@ const _options = Symbol('options');
 const _page = Symbol('page');
 const _parseOptions = Symbol('parseOptions');
 const _provider = Symbol('provider');
+const _roundDimension = Symbol('roundDimension');
+const _roundDimensions = Symbol('roundDimensions');
 const _setDimensions = Symbol('setDimensions');
 const _tempFile = Symbol('tempFile');
 const _validate = Symbol('validate');
@@ -103,7 +106,7 @@ class Converter {
    *
    * @param {Buffer|string} input - the SVG input to be converted to another format
    * @param {Converter~ConvertOptions} [options] - the options to be used
-   * @return {Promise.<Buffer, Error>} A <code>Promise</code> that is resolved with the converted output buffer.
+   * @return {Promise<Buffer>} A <code>Promise</code> that is resolved with the converted output buffer.
    * @public
    */
   async convert(input, options) {
@@ -136,7 +139,7 @@ class Converter {
    *
    * @param {string} inputFilePath - the path of the SVG file to be converted to another file format
    * @param {Converter~ConvertFileOptions} [options] - the options to be used
-   * @return {Promise.<string, Error>} A <code>Promise</code> that is resolved with the output file path.
+   * @return {Promise<string>} A <code>Promise</code> that is resolved with the output file path.
    * @public
    */
   async convertFile(inputFilePath, options) {
@@ -162,7 +165,7 @@ class Converter {
    *
    * An error will occur if any problem arises while closing the browser, where applicable.
    *
-   * @return {Promise.<void, Error>} A <code>Promise</code> that is resolved once this {@link Converter} has been
+   * @return {Promise<void>} A <code>Promise</code> that is resolved once this {@link Converter} has been
    * destroyed.
    * @public
    */
@@ -188,28 +191,31 @@ class Converter {
     input = Buffer.isBuffer(input) ? input.toString('utf8') : input;
 
     const { provider } = this;
-    const start = input.indexOf('<svg');
+    const svg = cheerio.html(cheerio.load(input, null, false)('svg'));
 
-    let html = `<!DOCTYPE html>
+    let html = '';
+    if (svg) {
+      html += `<!DOCTYPE html>
+<html>
+<head>
 <base href="${options.baseUrl}">
+<meta charset="utf-8">
 <style>
 * { margin: 0; padding: 0; }
 html { background-color: ${provider.getBackgroundColor(options)}; }
-</style>`;
-    if (start >= 0) {
-      html += input.substring(start);
+</style>
+</head>
+<body>${svg}</body>
+</html>`;
     } else {
-      throw new Error('SVG element open tag not found in input. Check the SVG input');
+      throw new Error('SVG element not found in input. Check the SVG input');
     }
 
     const page = await this[_getPage](html);
 
     await this[_setDimensions](page, options);
 
-    const dimensions = await this[_getDimensions](page);
-    if (!dimensions) {
-      throw new Error('Unable to derive width and height from SVG. Consider specifying corresponding options');
-    }
+    const dimensions = await this[_getDimensions](page, options);
 
     if (options.scale !== 1) {
       dimensions.height *= options.scale;
@@ -218,10 +224,7 @@ html { background-color: ${provider.getBackgroundColor(options)}; }
       await this[_setDimensions](page, dimensions);
     }
 
-    await page.setViewport({
-      height: Math.round(dimensions.height),
-      width: Math.round(dimensions.width)
-    });
+    await page.setViewport(dimensions);
 
     const output = await page.screenshot(Object.assign({
       type: provider.getType(),
@@ -231,17 +234,33 @@ html { background-color: ${provider.getBackgroundColor(options)}; }
     return output;
   }
 
-  [_getDimensions](page) {
-    return page.evaluate(() => {
+  async [_getDimensions](page, options) {
+    const dimensions = await page.evaluate(() => {
       const el = document.querySelector('svg');
       if (!el) {
         return null;
       }
 
-      const widthIsPercent = (el.getAttribute('width') || '').endsWith('%');
-      const heightIsPercent = (el.getAttribute('height') || '').endsWith('%');
-      const width = !widthIsPercent && parseFloat(el.getAttribute('width'));
-      const height = !heightIsPercent && parseFloat(el.getAttribute('height'));
+      function parseAttributeDimension(attributeName) {
+        const attributeValue = el.getAttribute(attributeName);
+        if (!attributeValue || attributeValue.endsWith('%')) {
+          return null;
+        }
+
+        const dimension = parseFloat(attributeValue);
+        if (Number.isNaN(dimension)) {
+          return null;
+        }
+
+        if (attributeValue.endsWith('pt')) {
+          return dimension * 1.33333;
+        }
+
+        return dimension;
+      }
+
+      const width = parseAttributeDimension('width');
+      const height = parseAttributeDimension('height');
 
       if (width && height) {
         return { width, height };
@@ -266,6 +285,11 @@ html { background-color: ${provider.getBackgroundColor(options)}; }
 
       return null;
     });
+    if (!dimensions) {
+      throw new Error('Unable to derive width and height from SVG. Consider specifying corresponding options');
+    }
+
+    return this[_roundDimensions](dimensions, options.rounding);
   }
 
   async [_getPage](html) {
@@ -328,9 +352,15 @@ html { background-color: ${provider.getBackgroundColor(options)}; }
     if (typeof options.height === 'string') {
       options.height = parseInt(options.height, 10);
     }
+
+    if (typeof options.rounding !== 'string' || !['ceil', 'floor', 'round'].includes(options.rounding)) {
+      options.rounding = 'round';
+    }
+
     if (options.scale == null) {
       options.scale = 1;
     }
+
     if (typeof options.width === 'string') {
       options.width = parseInt(options.width, 10);
     }
@@ -338,6 +368,25 @@ html { background-color: ${provider.getBackgroundColor(options)}; }
     provider.parseAPIOptions(options, inputFilePath);
 
     return options;
+  }
+
+  [_roundDimension](dimension, rounding) {
+    switch (rounding) {
+      case 'ceil':
+        return Math.ceil(dimension);
+      case 'floor':
+        return Math.floor(dimension);
+      case 'round':
+      default:
+        return Math.round(dimension);
+    }
+  }
+
+  [_roundDimensions](dimensions, rounding) {
+    return {
+      width: this[_roundDimension](dimensions.width, rounding),
+      height: this[_roundDimension](dimensions.height, rounding),
+    };
   }
 
   async [_setDimensions](page, dimensions) {
@@ -416,6 +465,8 @@ module.exports = Converter;
  * conjunction with the <code>baseFile</code> option.
  * @property {number|string} [height] - The height of the output to be generated. If omitted, an attempt will be made to
  * derive the height from the SVG input.
+ * @property {Converter~Rounding} [rounding] - The type of rounding to be applied to the width and height. If omitted,
+ * the dimensions with be rounded to the nearest integer.
  * @property {number} [scale=1] - The scale to be applied to the width and height (either specified as options or
  * derived).
  * @property {number|string} [width] - The width of the output to be generated. If omitted, an attempt will be made to
@@ -428,4 +479,10 @@ module.exports = Converter;
  * @typedef {Object} Converter~Options
  * @property {Object} [puppeteer] - The options that are to be passed directly to <code>puppeteer.launch</code> when
  * creating the <code>Browser</code> instance.
+ */
+
+/**
+ * The type of rounding to be applied to the width and height during a conversion.
+ *
+ * @typedef {'ceil'|'floor'|'round'} Converter~Rounding
  */